Allow canteen_id change on PATCH
My first idea was that the canteen_id represents the favs 'identity' and allowing to
change it after creating seemed odd (workaround: delete this one, create a new one).
But from server's perspective the canteen_id is a value that has to fit some validations
(canteen exists etc.) so we should allow this.